arm64: dts: qcom: Import Nubia Mini 5G (TP1803) device-tree overlay
The device-tree overlay was done manually under the following findings
where sm8150-v2.dts was recognized as the base DTB and sm8150-sdx50m-mtp
-overlay.dts as the base DTBO as per the Machine type printed in early
boot logs. This was cross-checked against compatible strings in FDT
(Flattened Device-Tree) and the userland device-tree.

The above-mentioned device-trees are mirrored and prefixed with "tp1803-"
as an identifier and added inclusions to the top-level overlaid DTB and
DTBO bits. The device-trees are named in such a manner that where the
root word is the device-tree that it was overlaid after.

The resulting layout of DTOs within QCOM devicetree source is as follows:

  qcom
  ├── tp1803
      ├── dtbo-overlay
      │   ├── include
      │   │   ├── battery
      │   │   │   ├── fg-gen4-batterydata-alium-3450mah-nubia-old.dtsi
      │   │   │   └── fg-gen4-batterydata-alium-3600mah-nubia.dtsi
      │   │   ├── display
      │   │   │   ├── nubia-jdi-hx83112a-1080-2160-5p65-video.dtsi
      │   │   │   └── nubia-lead-hx83112a-1080-2160-5p65-video.dtsi
      │   │   └── oem
      │   │       ├── nubia-display.dtsi
      │   │       ├── nubia-finger.dtsi
      │   │       ├── nubia-gpio_key.dtsi
      │   │       ├── nubia-head.dtsi
      │   │       ├── nubia-leds.dtsi
      │   │       ├── nubia-neo-power.dtsi
      │   │       ├── nubia-sensors.dtsi
      │   │       └── nubia-touch.dtsi
      │   ├── tp1803-base-overlay.dtsi
      │   ├── tp1803-pm8150b.dtsi
      │   ├── tp1803-sm8150-audio-overlay.dtsi
      │   ├── tp1803-sm8150-camera-sensor-mtp.dtsi
      │   ├── tp1803-sm8150-mtp.dtsi
      │   └── tp1803-sm8150-sde-display.dtsi
      └── dtb-overlay
          ├── tp1803-base.dtsi
          ├── tp1803-msm-audio-lpass.dtsi
          ├── tp1803-sm8150-audio.dtsi
          ├── tp1803-sm8150-pcie.dtsi
          └── tp1803-sm8150-pinctrl.dtsi

tp1803/dtb-overlay directory contains the overlaid DTB bits where
tp1803-base.dtsi is at the top-level which includes the other bits.

tp1803/dtbo-overlay contains the overlaid DTBO bits where tp1803-base
-overlay.dtsi is at top-level which includes the other bits. There is
also an include folder with batterydata, display drivers and the bits
that were overlaid by the device OEM under the oem folder with fixed
inclusions pertaining to the new layout and miscellaneous changes,
these bits are prefixed with "nubia-" to provide a clear identifier.

Omit building other QCOM board DTBs and DTBOs to save space and compile
times, this should also potentially speed up boot as the bootloader will
have fewer to DTBOs to compare for overlaying.

Signed-off-by: Daniel Jacob Chittoor <djchittoor47@gmail.com>