hwmon: tmd2725: Disable verbose logging
This suppresses overly-verbose logging of tmd2725_get_lux function which
prints increments or decrements made during adaptive brightness update
events. Stub out info logging under this clause while we are at it, as
this printk wrapper is used for printing parameters, offsets and
instructions of the sensor's ALS and proximity during each and every
integration cycle.

Signed-off-by: Daniel Jacob Chittoor <djchittoor47@gmail.com>